// src/app/models/specs.ts
export enum Specs {
  BloodDk = 'Blood Death Knight',
  FrostDk = 'Frost Death Knight',
  UnholyDk = 'Unholy Death Knight',
  BalanceDruid = 'Balance Druid',
  FeralDruid = 'Feral Druid',
  GuardianDruid = 'Guardian Druid',
  RestoDruid = 'Restoration Druid',
  MarksmanshipHunter = 'Marksmanship Hunter',
  SurvivalHunter = 'Survival Hunter',
  ArcaneMage = 'Arcane Mage',
  FireMage = 'Fire Mage',
  FrostMage = 'Frost Mage',
  HolyPaladin = 'Holy Paladin',
  ProtectionPaladin = 'Protection Paladin',
  RetributionPaladin = 'Retribution Paladin',
  DisciplinePriest = 'Discipline Priest',
  HolyPriest = 'Holy Priest',
  ShadowPriest = 'Shadow Priest',
  AssassinationRogue = 'Assassination Rogue',
  OutlawRogue = 'Outlaw Rogue',
  SubtletyRogue = 'Subtlety Rogue',
  ElementalShaman = 'Elemental Shaman',
  EnhancementShaman = 'Enhancement Shaman',
  RestorationShaman = 'Restoration Shaman',
  AfflictionWarlock = 'Affliction Warlock',
  DemonologyWarlock = 'Demonology Warlock',
  DestructionWarlock = 'Destruction Warlock',
  ArmsWarrior = 'Arms Warrior',
  FuryWarrior = 'Fury Warrior',
  ProtectionWarrior = 'Protection Warrior',
  BrewmasterMonk = 'Brewmaster Monk',
  MistweaverMonk = 'Mistweaver Monk',
  WindwalkerMonk = 'Windwalker Monk',
  BeastMasteryHunter = 'Beast Mastery Hunter',
  HavocDH = 'Havoc Demon Hunter',
  VengeanceDH = 'Vengeance Demon Hunter',
}